Rename MarketWithAuction to Marketplace
