+++ artifacts/contracts/v2/Marketplace.sol/Marketplace.json
@@ -1,7 +1,7 @@
 {
   "_format": "hh-sol-artifact-1",
-  "contractName": "MarketWithAuction",
-  "sourceName": "contracts/v2/MarketWithAuction.sol",
+  "contractName": "Marketplace",
+  "sourceName": "contracts/v2/Marketplace.sol",
   "abi": [
     {
       "inputs": [
