+++ test/v2/Marketplace/createDirectListing.ts
@@ -4,7 +4,7 @@ import { solidity } from "ethereum-waffle";
 
 // Contract Types
 import { MockERC1155 } from "../../../typechain/MockERC1155";
-import { MarketWithAuction, ListingParametersStruct, ListingStruct } from "../../../typechain/MarketWithAuction";
+import { Marketplace, ListingParametersStruct, ListingStruct } from "../../../typechain/Marketplace";
 
 // Types
 import { BigNumber } from "ethers";
@@ -15,7 +15,8 @@ import { getContracts, Contracts } from "../../../utils/tests/getContracts";
 
 use(solidity);
 
-describe("List token for sale: Auction Listing", function () {
+describe("List token for sale: Direct Listing", function () {
+
   // Constants
   const NATIVE_TOKEN_ADDRESS: string = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
 
@@ -25,7 +26,7 @@ describe("List token for sale: Auction Listing", function () {
   let lister: SignerWithAddress;
 
   // Contracts
-  let marketv2: MarketWithAuction;
+  let marketv2: Marketplace;
   let mockNft: MockERC1155;
 
   // MockERC1155: `mint` parameters
@@ -87,14 +88,14 @@ describe("List token for sale: Auction Listing", function () {
       reservePricePerToken: ethers.utils.parseEther("0.1"),
       buyoutPricePerToken: ethers.utils.parseEther("0.2"),
 
-      listingType: ListingType.Auction
+      listingType: ListingType.Direct
     }
   });
 
   describe("Revert cases", function() {
     it("Should revert if listing zero quantity.", async () => {
-      
-      // Invalid behaviour: listing 0 NFTs for auction.
+
+      // Invalid behaviour: listing `0` of the tokens for sale.
       const incorrectParams = {...listingParams, quantityToList: 0 };
 
       await expect(
@@ -102,9 +103,9 @@ describe("List token for sale: Auction Listing", function () {
       ).to.be.revertedWith("Market: must own and approve to transfer tokens.")
     })
 
-    it("Should revert if lister has insufficient token balance", async () => {
-
-      // Invalid behaviour: lister does not own the amount of tokens they they want to auction.
+    it("Should revert if lister does not own the amount of token to list.", async () => {
+      
+      // Invalid behaviour: caller does not own the given amount of tokens being listed.
       await expect(
         marketv2.connect(protocolAdmin).createListing(listingParams)
       ).to.be.revertedWith("Market: must own and approve to transfer tokens.")
@@ -112,7 +113,7 @@ describe("List token for sale: Auction Listing", function () {
 
     it("Should revert if market doesn't have lister's approval to transfer tokens", async () => {
 
-      // Invalid behaviour: `lister` has not approved Market to transfer tokens being listed.
+      // Invalid behaviour: `lister` does not approve Market to transfer tokens listed.
       await approveMarketToTransferTokens(false);
 
       await expect(
@@ -124,15 +125,15 @@ describe("List token for sale: Auction Listing", function () {
   describe("Events", function() {
 
     it("Should emit NewListing with all relevant info", async () => {
-    
+
       /**
        * Hardhat increments block timestamp by 1 on every transaction.
        * So, the timestamp during the `createListing` transaction will be the current timestamp + 1.
-       */
+      */
       // const timeStampOnCreateListing: BigNumber = BigNumber.from((await ethers.provider.getBlock("latest")).timestamp + 1);
 
       await expect(
-        marketv2.connect(lister).createListing(listingParams)  
+        marketv2.connect(lister).createListing(listingParams)
       ).to.emit(marketv2, "NewListing")
       .withArgs(
         mockNft.address,
@@ -150,25 +151,22 @@ describe("List token for sale: Auction Listing", function () {
           reservePricePerToken: listingParams.reservePricePerToken,
           buyoutPricePerToken: listingParams.buyoutPricePerToken,
           tokenType: TokenType.ERC1155,
-          listingType: ListingType.Auction
+          listingType: ListingType.Direct
         })
-      )
+      );
+      
+      
     })
   })
 
   describe("Token balances", function() {
 
-    it("Should transfer tokens listed from lister to market", async () => {
-      const listerBalBefore: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
-      const marketBalBefore: BigNumber = await mockNft.balanceOf(marketv2.address, nftTokenId);
-
+    it("Should not change the lister's token balance", async () => {
+      const balBefore: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
       await marketv2.connect(lister).createListing(listingParams);
+      const balAfter: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
       
-      const listerBalAfter: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
-      const marketBalAfter: BigNumber = await mockNft.balanceOf(marketv2.address, nftTokenId);
-
-      expect(listerBalBefore).to.equal(listerBalAfter.add(listingParams.quantityToList))
-      expect(marketBalAfter).to.equal(marketBalBefore.add(listingParams.quantityToList))
+      expect(balBefore).to.equal(balAfter)
     })
   })
 
@@ -183,7 +181,7 @@ describe("List token for sale: Auction Listing", function () {
     })
 
     it("Should store the listing's state", async () => {
-
+      const listingId: BigNumber = await marketv2.totalListings();
       const timeStampOnCreateListing: BigNumber = BigNumber.from((await ethers.provider.getBlock("latest")).timestamp + 1);
 
       await marketv2.connect(lister).createListing(listingParams);
@@ -201,7 +199,7 @@ describe("List token for sale: Auction Listing", function () {
       expect(listing.reservePricePerToken).to.equal(listingParams.reservePricePerToken);
       expect(listing.buyoutPricePerToken).to.equal(listingParams.buyoutPricePerToken);
       expect(listing.tokenType).to.equal(TokenType.ERC1155)
-      expect(listing.listingType).to.equal(ListingType.Auction);
+      expect(listing.listingType).to.equal(ListingType.Direct);
     })
   })
 });
