+++ test/v2/Marketplace/offerAuctionListing/withNativeToken.ts
@@ -4,8 +4,8 @@ import { solidity } from "ethereum-waffle";
 
 // Contract Types
 import { MockERC1155 } from "../../../../typechain/MockERC1155";
-import { Coin } from "../../../../typechain/Coin";
-import { MarketWithAuction, ListingParametersStruct, ListingStruct } from "../../../../typechain/MarketWithAuction";
+import { WETH9 } from "../../../../typechain/WETH9";
+import { Marketplace, ListingParametersStruct, ListingStruct } from "../../../../typechain/Marketplace";
 
 // Types
 import { BigNumber } from "ethers";
@@ -16,7 +16,10 @@ import { getContracts, Contracts } from "../../../../utils/tests/getContracts";
 
 use(solidity);
 
-describe("Bid with ERC20 token: Auction Listing", function () {
+describe("Bid with native token: Auction Listing", function () {
+  // Constants
+  const NATIVE_TOKEN_ADDRESS: string = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
+
   // Signers
   let protocolProvider: SignerWithAddress;
   let protocolAdmin: SignerWithAddress;
@@ -25,9 +28,9 @@ describe("Bid with ERC20 token: Auction Listing", function () {
   let dummy: SignerWithAddress;
 
   // Contracts
-  let marketv2: MarketWithAuction;
+  let marketv2: Marketplace;
   let mockNft: MockERC1155;
-  let erc20Token: Coin;
+  let weth: WETH9;
 
   // MockERC1155: `mint` parameters
   const nftTokenId: BigNumber = BigNumber.from(1);
@@ -55,14 +58,6 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     ethers.utils.toUtf8Bytes("")
   );
 
-  const mintERC20To = async (to: SignerWithAddress, amount: BigNumber) => {
-    // Mint currency to buyer
-    await erc20Token.connect(protocolAdmin).mint(to.address, amount);
-
-    // Approve Market to transfer currency
-    await erc20Token.connect(to).approve(marketv2.address, amount);
-  }
-
   const approveMarketToTransferTokens = async (toApprove: boolean) => await mockNft.connect(lister).setApprovalForAll(marketv2.address, toApprove);
   
   const timeTravelToListingWindow = async (listingId: BigNumber) => {
@@ -88,7 +83,7 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     mockNft = await ethers.getContractFactory("MockERC1155").then(f => f.connect(protocolAdmin).deploy());
     const contracts: Contracts = await getContracts(protocolProvider, protocolAdmin);
     marketv2 = contracts.marketv2;
-    erc20Token = contracts.coin;
+    weth = contracts.weth;
 
     // Setup: mint NFT to `lister` for `lister` to list these NFTs for sale.
     await mintNftToLister();
@@ -110,7 +105,7 @@ describe("Bid with ERC20 token: Auction Listing", function () {
       secondsUntilEndTime: BigNumber.from(1000),
 
       quantityToList: nftTokenSupply,
-      currencyToAccept: erc20Token.address,
+      currencyToAccept: NATIVE_TOKEN_ADDRESS,
 
       reservePricePerToken: ethers.utils.parseEther("0.01"),
       buyoutPricePerToken: ethers.utils.parseEther("0.02"),
@@ -126,12 +121,6 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     offerPricePerToken = listingParams.reservePricePerToken as BigNumber;
     currencyForOffer = listingParams.currencyToAccept;
     totalOfferAmount = offerPricePerToken.mul(listingParams.quantityToList);
-
-    // Setup: mint some curreny to buyer so they can fulfill the offer made.
-    await mintERC20To(
-      buyer,
-      (listingParams.buyoutPricePerToken as BigNumber).mul(listingParams.quantityToList)
-    )
   });
   
   describe("Revert cases", async () => {
@@ -141,49 +130,31 @@ describe("Bid with ERC20 token: Auction Listing", function () {
       await timeTravelToListingWindow(listingId);
       
       // Invalid behaviour: total offer amount is less than reserve price per token * quantity of auctioned item.
-      const invalidOfferPricePerToken = (listingParams.reservePricePerToken as BigNumber).sub(
-        ethers.utils.parseEther("0.005")
-      )
+      const invalidOfferAmount = (offerPricePerToken.sub(ethers.utils.parseEther("0.005"))).mul(quantityWanted);
 
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, invalidOfferPricePerToken)
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: invalidOfferAmount })
       ).to.be.revertedWith("Market: must offer at least reserve price.")
     })
 
-    it("Should revert if bid is made outside the auction window", async () => {      
+    it("Should revert if bid is made outside the auction window", async () => {
 
       // Invalid behaviour: bid is made outside auction window.
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount })
       ).to.be.revertedWith("Market: can only make offers in listing duration.")
 
       await timeTravelToAfterListingWindow(listingId)
 
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken),
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount }),
       ).to.be.revertedWith("Market: can only make offers in listing duration.")
     })
 
-    it("Should revert if buyer does not own the required amount of currency", async () => {
-      
-      // Invalid behaviour: buyer does not own the offer amount
-      await erc20Token.connect(buyer).transfer(protocolAdmin.address, await erc20Token.balanceOf(buyer.address));
+    it("Should revert if offer amount does not match native token sent with transaction.", async () => {
 
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
-      ).to.be.reverted;
-    })
-
-    it("Should revert if buyer has not approved Market to transfer currency", async () => {
-
-      // Invalid behaviour: buyer has not approved Market to transfer currency.
-      await erc20Token.connect(buyer).decreaseAllowance(
-        marketv2.address,
-        await erc20Token.allowance(buyer.address, marketv2.address)
-      );
-
-      await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: 0 })
       ).to.be.reverted;
     })
   })
@@ -199,7 +170,7 @@ describe("Bid with ERC20 token: Auction Listing", function () {
       const listing: ListingStruct = await marketv2.listings(listingId);
 
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount })
       ).to.emit(marketv2, "NewOffer")
       .withArgs(
         listingId,
@@ -229,11 +200,11 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     })
 
     it("Should emit NewBid if the incoming bid is the new highest bid, but below buyout price", async () => {
-        
+
       const listing: ListingStruct = await marketv2.listings(listingId);
 
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken) 
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount }) 
       ).to.emit(marketv2, "NewBid")
       .withArgs(
         listingId,
@@ -268,12 +239,12 @@ describe("Bid with ERC20 token: Auction Listing", function () {
         .mul(listingParams.quantityToList)
         .sub(ethers.utils.parseEther("0.001")); // We don't want the auction to close
       
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, highOfferAmount.div(listingParams.quantityToList))
+      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, highOfferAmount.div(listingParams.quantityToList), { value: highOfferAmount })
 
-      const lowOfferPricePerToken = offerPricePerToken;
+      const lowOfferAmount = totalOfferAmount;
 
       let secondBuyer = protocolAdmin // doesn't matter who
-      const txReceipt = await marketv2.connect(secondBuyer).offer(listingId, quantityWanted, currencyForOffer, lowOfferPricePerToken)
+      const txReceipt = await marketv2.connect(secondBuyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: lowOfferAmount })
         .then(tx => tx.wait());
       
       const newBidTopic = marketv2.interface.getEventTopic("NewBid");
@@ -283,11 +254,13 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     })
     
     it("Should emit AuctionClosed if bid is greater or equal to buyout price", async () => {
+      
+      const buyoutOfferAmount = (listingParams.buyoutPricePerToken as BigNumber).mul(listingParams.quantityToList);
       const listing: ListingStruct = await marketv2.listings(listingId);
       const timeStampOfBuyout = (await ethers.provider.getBlock("latest")).timestamp + 1;
 
       await expect(
-        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, listingParams.buyoutPricePerToken)
+        marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, listingParams.buyoutPricePerToken, { value: buyoutOfferAmount })
       ).to.emit(marketv2, "AuctionClosed")
       .withArgs(
         listingId,
@@ -326,56 +299,63 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     })
     
     it("Should escrow currency in Market if bid is valid", async () => {
-      const buyerBalBefore: BigNumber = await erc20Token.balanceOf(buyer.address)
-      const marketBalBefore: BigNumber = await erc20Token.balanceOf(marketv2.address);
 
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
+      const buyerBalBefore: BigNumber = await ethers.provider.getBalance(buyer.address)
+      const marketBalBefore: BigNumber = await weth.balanceOf(marketv2.address);
 
-      const buyerBalAfter: BigNumber = await erc20Token.balanceOf(buyer.address)
-      const marketBalAfter: BigNumber = await erc20Token.balanceOf(marketv2.address);
+      const gasPrice = ethers.utils.parseUnits("1", "gwei");
+      const tx = await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount, gasPrice })
+      const gasUsed = (await tx.wait()).gasUsed;
+      const gasPaid = gasPrice.mul(gasUsed);
 
-      expect(buyerBalAfter).to.equal(buyerBalBefore.sub(totalOfferAmount))
+      const buyerBalAfter: BigNumber = await ethers.provider.getBalance(buyer.address)
+      const marketBalAfter: BigNumber = await weth.balanceOf(marketv2.address);
+
+      expect(buyerBalAfter).to.equal(buyerBalBefore.sub(totalOfferAmount.add(gasPaid)))
       expect(marketBalAfter).to.equal(marketBalBefore.add(totalOfferAmount));
     })
 
     it("Should payout the previous bidder, if new bid is higher than the previous bid", async () => {
-      const highOfferAmount = (listingParams.reservePricePerToken as BigNumber)
-        .mul(listingParams.quantityToList)
-        .add(ethers.utils.parseEther("0.01"));
+      const highOfferAmount = (offerPricePerToken.mul(listingParams.quantityToList))
+        .add(ethers.utils.parseEther("0.5"));
 
-      const lowOfferPricePerToken = offerPricePerToken;
+      const lowOfferAmount = totalOfferAmount;
 
-      // Set up `prevBuyer`
       const prevBuyer = protocolAdmin; // doesn't matter who
-      await mintERC20To(
-        prevBuyer,
-        highOfferAmount
-      )
       
-      const prevBuyerBalBefore: BigNumber = await erc20Token.balanceOf(prevBuyer.address)
-      const buyerBalBefore: BigNumber = await erc20Token.balanceOf(buyer.address)
-      const marketBalBefore: BigNumber = await erc20Token.balanceOf(marketv2.address);
+      const prevBuyerBalBefore: BigNumber = await ethers.provider.getBalance(prevBuyer.address)
+      const buyerBalBefore: BigNumber = await ethers.provider.getBalance(buyer.address)
+      const marketBalBefore: BigNumber = await weth.balanceOf(marketv2.address);
 
-      await marketv2.connect(prevBuyer).offer(listingId, quantityWanted, currencyForOffer, lowOfferPricePerToken);
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, highOfferAmount.div(listingParams.quantityToList));
-
-      const prevBuyerBalAfter: BigNumber = await erc20Token.balanceOf(prevBuyer.address)
-      const buyerBalAfter: BigNumber = await erc20Token.balanceOf(buyer.address)
-      const marketBalAfter: BigNumber = await erc20Token.balanceOf(marketv2.address);
+      const gasPrice = ethers.utils.parseUnits("1", "gwei");
       
-      expect(prevBuyerBalAfter).to.equal(prevBuyerBalBefore)
-      expect(buyerBalAfter).to.equal(buyerBalBefore.sub(highOfferAmount))
+      const tx1 = await marketv2.connect(prevBuyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: lowOfferAmount, gasPrice });
+      const tx2 = await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, highOfferAmount.div(listingParams.quantityToList), { value: highOfferAmount, gasPrice });
+
+      const gasUsed1 = (await tx1.wait()).gasUsed;
+      const gasUsed2 = (await tx2.wait()).gasUsed;
+      const gasPaid1 = gasPrice.mul(gasUsed1);
+      const gasPaid2 = gasPrice.mul(gasUsed2);
+      
+      const prevBuyerBalAfter: BigNumber = await ethers.provider.getBalance(prevBuyer.address)
+      const buyerBalAfter: BigNumber = await ethers.provider.getBalance(buyer.address)
+      const marketBalAfter: BigNumber = await weth.balanceOf(marketv2.address);
+      
+      expect(prevBuyerBalAfter).to.equal(prevBuyerBalBefore.sub(gasPaid1))
+      expect(buyerBalAfter).to.equal(buyerBalBefore.sub(highOfferAmount.add(gasPaid2)))
       expect(marketBalAfter).to.equal(marketBalBefore.add(highOfferAmount));
     });
 
     it("Should transfer auctioned tokens to bidder if bid is at buyout price.", async () => {
+      
+      const buyoutOfferAmount = (listingParams.buyoutPricePerToken as BigNumber).mul(listingParams.quantityToList);
 
       const auctionQuantity: BigNumber =  (await marketv2.listings(listingId)).quantity;
 
       const buyerBalBefore: BigNumber = await mockNft.balanceOf(buyer.address, nftTokenId)
       const marketBalBefore: BigNumber = await mockNft.balanceOf(marketv2.address, nftTokenId);
 
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, listingParams.buyoutPricePerToken)
+      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, listingParams.buyoutPricePerToken, { value: buyoutOfferAmount })
 
       const buyerBalAfter: BigNumber = await mockNft.balanceOf(buyer.address, nftTokenId)
       const marketBalAfter: BigNumber = await mockNft.balanceOf(marketv2.address, nftTokenId);
@@ -391,7 +371,8 @@ describe("Bid with ERC20 token: Auction Listing", function () {
     })
 
     it("Should store a valid offer regardless", async () => {
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
+
+      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount })
 
       const offer = await marketv2.offers(listingId, buyer.address);
 
@@ -401,8 +382,9 @@ describe("Bid with ERC20 token: Auction Listing", function () {
       expect(offer.pricePerToken).to.equal(offerPricePerToken);
     })
 
-    it("Should store the offer as the winning bid if it is the new highest bid", async () => {
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken)
+    it("Should store the offer as the winning bid if it is the new highest bid", async () => {      
+
+      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount })
 
       const winningBid = await marketv2.winningBid(listingId);
 
@@ -419,7 +401,7 @@ describe("Bid with ERC20 token: Auction Listing", function () {
 
       await ethers.provider.send("evm_mine", [endTimeBefore.sub(timeBuffer).add(1).toNumber()]);
 
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken);
+      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, offerPricePerToken, { value: totalOfferAmount });
 
       const endTimeAfter: BigNumber = (await marketv2.listings(listingId)).endTime;
 
@@ -428,7 +410,9 @@ describe("Bid with ERC20 token: Auction Listing", function () {
 
     it("Should close the auction by updating the listing's end time, if the bid is buyout price", async () => {
 
-      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, listingParams.buyoutPricePerToken);
+      const buyoutOfferAmount = (listingParams.buyoutPricePerToken as BigNumber).mul(listingParams.quantityToList);
+
+      await marketv2.connect(buyer).offer(listingId, quantityWanted, currencyForOffer, listingParams.buyoutPricePerToken, { value: buyoutOfferAmount });
 
       const timeStamp = (await ethers.provider.getBlock("latest")).timestamp;
       const endTimeAfter: BigNumber = (await marketv2.listings(listingId)).endTime;
